+++ src/index.ts
@@ -1,6 +1,23 @@
 import { Storage } from "./Storage";
 // import { AbstractAdapter } from "./AbstractAdapter";
 // import { parseUrl } from "./util";
-import { StorageType, IStorage, AdapterConfig } from "./types";
+import {
+  StorageType,
+  IStorage,
+  AdapterConfig,
+  ConfigAmazonS3,
+  ConfigBackblazeB2,
+  ConfigGoogleCloud,
+  ConfigLocal,
+} from "./types";
 
-export { Storage, IStorage, StorageType, AdapterConfig };
+export {
+  Storage,
+  IStorage,
+  StorageType,
+  AdapterConfig,
+  ConfigAmazonS3,
+  ConfigBackblazeB2,
+  ConfigGoogleCloud,
+  ConfigLocal,
+};